refactor(protocol): split frame module and start hardening
