+++ code/shared/bpane-protocol/src/frame/error.rs
@@ -1,0 +1,39 @@
+use core::fmt;
+
+/// Errors during frame encoding/decoding.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum FrameError {
+    /// Not enough data to read the expected field.
+    BufferTooShort { expected: usize, available: usize },
+    /// Unknown channel ID.
+    UnknownChannel(u8),
+    /// Unknown message type tag within a channel.
+    UnknownMessageType { channel: u8, tag: u8 },
+    /// Payload exceeds maximum allowed size.
+    PayloadTooLarge(u32),
+    /// Data remaining after parsing.
+    TrailingData(usize),
+}
+
+impl fmt::Display for FrameError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::BufferTooShort {
+                expected,
+                available,
+            } => write!(f, "buffer too short: need {expected}, have {available}"),
+            Self::UnknownChannel(ch) => write!(f, "unknown channel: 0x{ch:02x}"),
+            Self::UnknownMessageType { channel, tag } => {
+                write!(
+                    f,
+                    "unknown message type 0x{tag:02x} on channel 0x{channel:02x}"
+                )
+            }
+            Self::PayloadTooLarge(size) => write!(f, "payload too large: {size} bytes"),
+            Self::TrailingData(n) => write!(f, "{n} trailing bytes after message"),
+        }
+    }
+}
+
+#[cfg(feature = "std")]
+impl std::error::Error for FrameError {}
