+++ code/shared/bpane-protocol/src/frame/tests.rs
@@ -1,0 +1,9 @@
+mod audio;
+mod control;
+mod cursor_clipboard_file;
+mod envelope;
+mod input;
+mod message_dispatch;
+mod tile_frame;
+mod tile_roundtrip;
+mod video;
